+++ tests/integration_tests.rs
@@ -1,5 +1,7 @@
 use rand::RngExt;
 use rand::prelude::SliceRandom;
+#[cfg(feature = "experimental-api-5")]
+use redb::KeyRange;
 use redb::backends::FileBackend;
 use redb::{
     AccessGuard, Builder, CommitError, CompactionError, Database, Durability, Key, MultimapRange,
@@ -12,6 +14,7 @@ use std::borrow::Borrow;
 use std::fs;
 use std::io::{ErrorKind, Write};
 use std::marker::PhantomData;
+#[cfg(not(feature = "experimental-api-5"))]
 use std::ops::RangeBounds;
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::{Arc, Condvar, Mutex, RwLock, mpsc};
@@ -3879,6 +3882,12 @@ impl<K: Key + 'static, V: Value + 'static, T: ReadableTable<K, V>> ReadableTable
         self.inner.get(key)
     }
 
+    #[cfg(feature = "experimental-api-5")]
+    fn range<'a>(&self, range: impl KeyRange<'a, K>) -> redb::Result<Range<'_, K, V>> {
+        self.inner.range(range)
+    }
+
+    #[cfg(not(feature = "experimental-api-5"))]
     fn range<'a, KR>(&self, range: impl RangeBounds<KR> + 'a) -> redb::Result<Range<'_, K, V>>
     where
         KR: Borrow<K::SelfType<'a>> + 'a,
@@ -3936,6 +3945,12 @@ impl<K: Key + 'static, V: Key + 'static, T: ReadableMultimapTable<K, V>> Readabl
         self.inner.get(key)
     }
 
+    #[cfg(feature = "experimental-api-5")]
+    fn range<'a>(&self, range: impl KeyRange<'a, K>) -> redb::Result<MultimapRange<'_, K, V>> {
+        self.inner.range(range)
+    }
+
+    #[cfg(not(feature = "experimental-api-5"))]
     fn range<'a, KR>(
         &self,
         range: impl RangeBounds<KR> + 'a,